Minor fix: Rename climates.py in tuya devices data
